subnet: Move etcd files into their own package.
Kube backend is in it's own package so etcdv2 should be too
This makes it easier to see
- what being tested and what's not
- What interfaces are etcd specific (e.g. registry)
